feat: moved EarnSection to app/components/View/Homepage/Sections
