+++ app/components/Views/Homepage/Sections/EarnSection/index.ts
